requirements.txt
pandas
requests
python-dotenv
pytest
pytest-mock
fastapi
uvicorn
structlog
sqlalchemy
psycopg2-binary
influxdb-client
websockets
scikit-learn
joblib
aiosqlite
pytest-asyncio
redis
httpx
kafka-python

prometheus-client
prometheus-fastapi-instrumentator
matplotlib

cachetools
boto3
optuna
